+++ include/boost/beast/websocket/detail/frame.hpp
@@ -13,11 +13,9 @@
 #include <boost/beast/core/buffer_traits.hpp>
 #include <boost/beast/websocket/error.hpp>
 #include <boost/beast/websocket/rfc6455.hpp>
-#include <boost/beast/websocket/detail/utf8_checker.hpp>
+#include <boost/beast/core/buffers_prefix.hpp>
 #include <boost/beast/core/flat_static_buffer.hpp>
-#include <boost/asio/buffer.hpp>
-#include <boost/assert.hpp>
-#include <boost/endian/conversion.hpp>
+#include <boost/beast/core/detail/buffers_pair.hpp>
 #include <cstdint>
 
 namespace boost {
@@ -89,159 +87,44 @@ is_control(opcode op)
     return op >= opcode::close;
 }
 
-inline
+BOOST_BEAST_DECL
 bool
-is_valid_close_code(std::uint16_t v)
-{
-    switch(v)
-    {
-    case close_code::normal:            // 1000
-    case close_code::going_away:        // 1001
-    case close_code::protocol_error:    // 1002
-    case close_code::unknown_data:      // 1003
-    case close_code::bad_payload:       // 1007
-    case close_code::policy_error:      // 1008
-    case close_code::too_big:           // 1009
-    case close_code::needs_extension:   // 1010
-    case close_code::internal_error:    // 1011
-    case close_code::service_restart:   // 1012
-    case close_code::try_again_later:   // 1013
-        return true;
-
-    // explicitly reserved
-    case close_code::reserved1:         // 1004
-    case close_code::no_status:         // 1005
-    case close_code::abnormal:          // 1006
-    case close_code::reserved2:         // 1014
-    case close_code::reserved3:         // 1015
-        return false;
-    }
-    // reserved
-    if(v >= 1016 && v <= 2999)
-        return false;
-    // not used
-    if(v <= 999)
-        return false;
-    return true;
-}
+is_valid_close_code(std::uint16_t v);
 
 //------------------------------------------------------------------------------
 
 // Write frame header to dynamic buffer
 //
-template<class DynamicBuffer>
+BOOST_BEAST_DECL
 void
-write(DynamicBuffer& db, frame_header const& fh)
-{
-    std::size_t n;
-    std::uint8_t b[14];
-    b[0] = (fh.fin ? 0x80 : 0x00) | static_cast<std::uint8_t>(fh.op);
-    if(fh.rsv1)
-        b[0] |= 0x40;
-    if(fh.rsv2)
-        b[0] |= 0x20;
-    if(fh.rsv3)
-        b[0] |= 0x10;
-    b[1] = fh.mask ? 0x80 : 0x00;
-    if(fh.len <= 125)
-    {
-        b[1] |= fh.len;
-        n = 2;
-    }
-    else if(fh.len <= 65535)
-    {
-        b[1] |= 126;
-        auto len_be = endian::native_to_big(
-            static_cast<std::uint16_t>(fh.len));
-        std::memcpy(&b[2], &len_be, sizeof(len_be));
-        n = 4;
-    }
-    else
-    {
-        b[1] |= 127;
-        auto len_be = endian::native_to_big(
-            static_cast<std::uint64_t>(fh.len));
-        std::memcpy(&b[2], &len_be, sizeof(len_be));
-        n = 10;
-    }
-    if(fh.mask)
-    {
-        auto key_le = endian::native_to_little(
-            static_cast<std::uint32_t>(fh.key));
-        std::memcpy(&b[n], &key_le, sizeof(key_le));
-        n += 4;
-    }
-    db.commit(net::buffer_copy(
-        db.prepare(n), net::buffer(b)));
-}
+write(flat_static_buffer_base& db, frame_header const& fh);
 
 // Read data from buffers
 // This is for ping and pong payloads
 //
-template<class Buffers>
+BOOST_BEAST_DECL
 void
-read_ping(ping_data& data, Buffers const& bs)
-{
-    BOOST_ASSERT(buffer_bytes(bs) <= data.max_size());
-    data.resize(buffer_bytes(bs));
-    net::buffer_copy(net::mutable_buffer{
-        data.data(), data.size()}, bs);
-}
+read_ping(
+    ping_data& data,
+    beast::buffers_prefix_view<beast::detail::buffers_pair<true>> const& bs);
 
 // Read close_reason, return true on success
 // This is for the close payload
 //
-template<class Buffers>
+BOOST_BEAST_DECL
 void
 read_close(
     close_reason& cr,
-    Buffers const& bs,
-    error_code& ec)
-{
-    auto const n = buffer_bytes(bs);
-    BOOST_ASSERT(n <= 125);
-    if(n == 0)
-    {
-        cr = close_reason{};
-        ec = {};
-        return;
-    }
-    if(n == 1)
-    {
-        // invalid payload size == 1
-        ec = error::bad_close_size;
-        return;
-    }
-
-    std::uint16_t code_be;
-    cr.reason.resize(n - 2);
-    std::array<net::mutable_buffer, 2> out_bufs{{
-        net::mutable_buffer(&code_be, sizeof(code_be)),
-        net::mutable_buffer(&cr.reason[0], n - 2)}};
-
-    net::buffer_copy(out_bufs, bs);
-
-    cr.code = endian::big_to_native(code_be);
-    if(! is_valid_close_code(cr.code))
-    {
-        // invalid close code
-        ec = error::bad_close_code;
-        return;
-    }
-
-    if(n > 2 && !check_utf8(
-        cr.reason.data(), cr.reason.size()))
-    {
-        // not valid utf-8
-        ec = error::bad_close_payload;
-        return;
-    }
-    ec = {};
-}
+    beast::buffers_prefix_view<beast::detail::buffers_pair<true>> const& bs,
+    error_code& ec);
 
 } // detail
 } // websocket
 } // beast
 } // boost
+
+#if BOOST_BEAST_HEADER_ONLY
+#include <boost/beast/websocket/detail/frame.ipp>
+#endif
 
 #endif