+++ include/boost/beast/websocket/impl/stream_impl.hpp
@@ -31,6 +31,7 @@
 #include <boost/beast/core/stream_traits.hpp>
 #include <boost/beast/core/detail/clamp.hpp>
 #include <boost/beast/version.hpp>
+#include <boost/endian/conversion.hpp>
 #include <boost/asio/steady_timer.hpp>
 #include <boost/core/empty_value.hpp>
 #include <boost/enable_shared_from_this.hpp>